changes in energy equations
The term dp/dt has been removed from dispersed phase equations.
The temperature of dispersed phase, when volume fraction is small, is set equal to that of continuous phase.
The correction  for small volume fraction in the temperature exchange term has been changed, not it always depends on the volumetric fraction of the dispersed phase.